+++ tk_qsrc/tk_clib/vxcore/tk_core/tk_mmpage.c
@@ -541,7 +541,7 @@ void *TKMM_PageAllocKrn(int sz)
 	ptr=TKMM_PageAlloc(sz);
 
 	TK_VMem_MProtectPages(
-		ptr, sz, TKMM_PROT_RW|TKMM_PROT_NOUSER);
+		(long)ptr, sz, TKMM_PROT_RW|TKMM_PROT_NOUSER);
 
 	return(ptr);
 }
